Updated GUI for Gas Data
Gas data from API is processed so that the Rs/Ro is compared in a look up table. When found the data is actual gas detected is obtained and so is the PPM. This is then displayed in the Faux-Iframe; .gas-display, .gas-info, .gas-paragraph, #PPM, .metadata

Added dashboard disclaimer to the footer of the page.